集成MyBatis Generator 完善User与Question相关操作

+++ src/main/java/com/community/controller/QuestionController.java
@@ -18,7 +18,7 @@ public class QuestionController {
     QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Long id, Model model) {
+    public String question(@PathVariable("id") Integer id, Model model) {
         if (id != null) {
             questionService.updateViewCount(id);
             QuestionDTO questionDTO = questionService.getById(id);
